requirements.txt
alembic==1.16.4
fastapi==0.116.1
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
SQLAlchemy==2.0.41
SQLAlchemy==2.0.36
uvicorn==0.35.0